Add #set_env step to Snapshot::Import
Rails 5 added some neat database protection in the form of a new table
called `ar_internal_metadata` used as a KV store. During normal
operation it stores the Rails env in the db upon the first run of
`db:migrate`. Since we're doing some fun workarounds for our snapshots,
this doesn't get set properly, and will throw
`ActiveRecord::NoEnvironmentInSchemaError` in the case that an
environment has not run `db:migrate` since the last
`db:snapshot:import`. This is most common in development but can present
in staging as well.

The solution is simply to run `db:environment:set`, so this commit adds
that as the final step to the `Snapshot::Import` service. This should
ensure that we never encounter this issue again.

Note that it is the final step because if we set the env at the
beginning of the task then we will truly have bypassed the Rails
protections and could potentially damage an env unintentionally. That
said, this gem also contains other safety measures that should avoid
that in case someone changes things without reading this.